(some refactoring)
