+++ packages/cli/src/lionweb-cli.ts
@@ -4,8 +4,8 @@
 import { argv } from "process"
 import { inferLanguages } from "./infer-languages-command.js"
 
-import { diagramFromSerializationChunkAt } from "./m3/diagram-command.js"
-import { generateTsTypesWith } from "./m3/generate-ts-types-command.js"
+import { diagramFromSerializationChunkAt } from "./m2/diagram-command.js"
+import { generateTsTypesWith } from "./m2/generate-ts-types-command.js"
 import { diffSerializationChunks } from "./serialization/diff-command.js"
 import { executeMeasureCommand } from "./serialization/measure-command.js"
 import { repairSerializationChunkAt } from "./serialization/repair-command.js"
@@ -145,7 +145,7 @@ textual syntax is used, unless a flag '--asRegular' is provided.
         }
 
         case VALIDATE_COMMAND: {
-            if (commandArgs.length === 0) {
+            if (commandArgs.length !== 1) {
                 console.log(
                     `The ${VALIDATE_COMMAND} command validates a serialization chunk.
 Usage: npx @lionweb/cli ${VALIDATE_COMMAND} <path_to_chunk>`
@@ -157,9 +157,10 @@ Usage: npx @lionweb/cli ${VALIDATE_COMMAND} <path_to_chunk>`
         }
 
         case INFER_LANGUAGES_COMMAND: {
-            if (commandArgs.length === 0) {
+            if (commandArgs.length !== 1) {
                 console.log(
-                    `The ${INFER_LANGUAGES_COMMAND} command infer language(s) from a given serialization chunk. \nUsage: npx @lionweb/cli infer-language <path_to_chunk>`
+                    `The ${INFER_LANGUAGES_COMMAND} command infer language(s) from a given serialization chunk.
+Usage: npx @lionweb/cli ${INFER_LANGUAGES_COMMAND} <path_to_chunk>`
                 )
             } else {
                 await inferLanguages(commandArgs[0])
